Fix type error shotline model and other small errors

+++ db/model/shotline.go
@@ -8,8 +8,8 @@ type Shotline struct {
 	Remark         string                 `json:"remark"`
 	ShotpointID    int32                  `json:"shotpoint_id"`
 	Number         float64                `json:"shotpoint_number"`
-	ShotpointMin   int16                  `json:"shotpoint_min"`
-	ShotpointMax   int16                  `json:"shotpoint_max"`
+	ShotpointMin   float64                `json:"shotpoint_min"`
+	ShotpointMax   float64                `json:"shotpoint_max"`
 	Stash          map[string]interface{} `json:"stash"`
 	KeywordSummary []KeywordSummary       `json:"keywords"`
 	GeoJSON        map[string]interface{} `json:"geojson"`
